Monitoring-Products/backend-app/requirements.txt
Django
django-cors-headers
mysqlclient
cryptography
mysql-connector-python
djangorestframework
requests